local business fetch moved to custom hook, with carousel and map refactored. helper function created to getTimeRemaining in carousel
